rename caddy/dal to dal/
