+++ go/caddy/api/dal_admin_api_audit_log_test.go
@@ -12,7 +12,7 @@ import (
 	"testing"
 
 	"github.com/inverse-inc/packetfence/go/admin_api_audit_log"
-	"github.com/inverse-inc/packetfence/go/caddy/dal/models"
+	"github.com/inverse-inc/packetfence/go/dal/models"
 	"github.com/inverse-inc/packetfence/go/db"
 	"github.com/julienschmidt/httprouter"
 	"gorm.io/driver/mysql"
